# CandidateService/src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Candidate Service API
  description: API for managing candidates and voting operations
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: Candidate API
    description: Operations related to candidates and voting

paths:

  /api/candidate/register:
    post:
      tags:
        - Candidate API
      summary: Register a new candidate
      description: Registers a candidate with initial vote count set to 0
      operationId: registerCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateRequestDTO'
      responses:
        '201':
          description: Candidate registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponseDTO'

  /api/candidate/getCandidateById/{id}:
    get:
      tags:
        - Candidate API
      summary: Get candidate by ID
      operationId: getCandidateById
      parameters:
        - name: id
          in: path
          required: true
          description: Candidate ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Candidate found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponseDTO'
        '404':
          description: Candidate not found

  /api/candidate/getAllCandidates:
    get:
      tags:
        - Candidate API
      summary: Get all candidates
      operationId: getAllCandidates
      responses:
        '200':
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidateResponseDTO'

  /api/candidate/updateCandidate/{id}:
    put:
      tags:
        - Candidate API
      summary: Update candidate
      operationId: updateCandidate
      parameters:
        - name: id
          in: path
          required: true
          description: Candidate ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateRequestDTO'
      responses:
        '202':
          description: Candidate updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateResponseDTO'

  /api/candidate/deleteCandidate/{id}:
    delete:
      tags:
        - Candidate API
      summary: Delete candidate
      operationId: deleteCandidate
      parameters:
        - name: id
          in: path
          required: true
          description: Candidate ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Candidate deleted successfully
          content:
            text/plain:
              schema:
                type: string

  /api/candidate/vote/{candidateId}:
    post:
      tags:
        - Candidate API
      summary: Vote for candidate
      operationId: voteForCandidate
      parameters:
        - name: candidateId
          in: path
          required: true
          description: Candidate ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Vote cast successfully
          content:
            text/plain:
              schema:
                type: string

  /api/candidate/removeVote/{candidateId}:
    post:
      tags:
        - Candidate API
      summary: Remove vote for candidate
      operationId: removeVoteForCandidate
      parameters:
        - name: candidateId
          in: path
          required: true
          description: Candidate ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Vote removed successfully
          content:
            text/plain:
              schema:
                type: string

  /api/candidate/electionResult:
    get:
      tags:
        - Candidate API
      summary: Get election result
      operationId: electionResult
      responses:
        '200':
          description: Election result string
          content:
            text/plain:
              schema:
                type: string

  /api/candidate/displayCandidateAndVoter:
    get:
      tags:
        - Candidate API
      summary: Display candidates and voters
      operationId: displayCandidateAndVoter
      responses:
        '200':
          description: List of candidates and voters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DisplayVotingDTO'

components:

  schemas:
    CandidateIdDTO:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          description: candidate id
          example: 123


    CandidateRequestDTO:
      type: object
      required:
        - id
        - candidateName
        - adharCardNo
        - party
      properties:
        id:
          type: integer
          format: int64
          example: 123
        candidateName:
          type: string
          example: Rahul Sharma
        adharCardNo:
          type: integer
          format: int64
          example: 123456789012
        party:
          type: string
          example: National Party

    CandidateResponseDTO:
      type: object
      properties:
        candidateId:
          type: integer
          format: int64
          example: 101
        candidateName:
          type: string
          example: Rahul Sharma
        adharCardNo:
          type: integer
          format: int64
          example: 123456789012
        party:
          type: string
          example: National Party

    DisplayVotingDTO:
      type: object
      properties:
        voterName:
          type: string
          example: Rahul Sharma
        hasVoted:
          type: string
          example: Yes
        candidateName:
          type: string
          example: Anita Desai
        party:
          type: string
          example: National Party

    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: john_doe
        password:
          type: string
          example: password123

    VoterClientResposeDTO:
      type: object
      properties:
        voterName:
          type: string
          example: Rahul Sharma
        hasVoted:
          type: string
          example: Yes